+++ src/peakfit/analysis/__init__.py
@@ -9,16 +9,6 @@ from peakfit.analysis.benchmarks import (
     create_synthetic_cluster,
     profile_fit_cluster,
 )
-from peakfit.analysis.caching import (
-    LRUCache,
-    cached_lstsq,
-    cached_shape_evaluation,
-    clear_shape_cache,
-    get_cache_stats,
-    hash_array,
-    make_cache_key,
-    memoize_array_function,
-)
 from peakfit.analysis.profiling import (
     Profiler,
     ProfileReport,
@@ -36,15 +26,6 @@ __all__ = [
     "compare_backends_report",
     "create_synthetic_cluster",
     "profile_fit_cluster",
-    # Caching
-    "LRUCache",
-    "cached_lstsq",
-    "cached_shape_evaluation",
-    "clear_shape_cache",
-    "get_cache_stats",
-    "hash_array",
-    "make_cache_key",
-    "memoize_array_function",
     # Profiling
     "Profiler",
     "ProfileReport",